# AvitoMerchShop/wwwroot/swagger/swagger.json
{
    "info": {
        "title": "API Avito shop",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/buy/{item}": {
            "get": {
                "summary": "Купить предмет за монеты.",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ."
                    },
                    "400": {
                        "description": "Неверный запрос.",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Неавторизован.",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера.",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "produces": [
                    "application/json"
                ]
            }
        },
...
    },
    "swagger": "2.0",
    "host": "localhost:8080",
    "schemes": [
        "http"
    ],
    "basePath": "/",
    "definitions": {
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "string",
                    "description": "Сообщение об ошибке, описывающее проблему."
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}